+++ packages/web-sdk/src/modules/region-picker/index.ts
@@ -1,11 +1,15 @@
 export type {
   AttachRegionPickerOptions,
+  BindRegionPickerOptions,
   RegionPicker,
   RegionPickerAttachment,
+  RegionPickerBinding,
+  RegionPickerElementIds,
   RegionPickerListener,
   RegionPickerOptions,
   RegionPickerState,
 } from './models';
 
 export { attachRegionPicker } from './attach-region-picker';
+export { bindRegionPicker } from './bind-region-picker';
 export { createRegionPicker } from './region-picker';